third_party/blink/web_tests/platform/generic/svg/animations/use-animate-transform-and-position-expected.txt

SVG 1.1 dynamic animation tests

Test animation of use element where the instance is also animated with animate and animateTransform

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.x.animVal.value is 10
PASS rect.x.baseVal.value is 10
PASS rect.y.animVal.value is 10
PASS rect.y.baseVal.value is 10
PASS use.x.animVal.value is 0
PASS use.x.baseVal.value is 0
PASS use.y.animVal.value is 0
PASS use.y.baseVal.value is 0
PASS use.transform.baseVal.numberOfItems is 0
PASS rect.x.animVal.value is 15
PASS rect.x.baseVal.value is 10
PASS rect.y.animVal.value is 10
PASS rect.y.baseVal.value is 10
PASS use.x.animVal.value is 0
PASS use.x.baseVal.value is 0
PASS use.y.animVal.value is 0
PASS use.y.baseVal.value is 0
PASS use.transform.animVal.numberOfItems is 1
PASS use.transform.baseVal.numberOfItems is 0
PASS use.transform.animVal.getItem(0).matrix.e is 0
PASS use.transform.animVal.getItem(0).matrix.f is 115
PASS rect.x.animVal.value is 25
PASS rect.x.baseVal.value is 10
PASS rect.y.animVal.value is 10
PASS rect.y.baseVal.value is 10
PASS use.x.animVal.value is 5
PASS use.x.baseVal.value is 0
PASS use.y.animVal.value is 0
PASS use.y.baseVal.value is 0
PASS use.transform.animVal.numberOfItems is 1
PASS use.transform.baseVal.numberOfItems is 0
PASS use.transform.animVal.getItem(0).matrix.e is 0
PASS use.transform.animVal.getItem(0).matrix.f is 125
PASS rect.x.animVal.value is 35
PASS rect.x.baseVal.value is 10
PASS rect.y.animVal.value is 10
PASS rect.y.baseVal.value is 10
PASS use.x.animVal.value is 15
PASS use.x.baseVal.value is 0
PASS use.y.animVal.value is 0
PASS use.y.baseVal.value is 0
PASS use.transform.animVal.numberOfItems is 1
PASS use.transform.baseVal.numberOfItems is 0
PASS use.transform.animVal.getItem(0).matrix.e is 0
PASS use.transform.animVal.getItem(0).matrix.f is 128.5
PASS rect.x.animVal.value is 45
PASS rect.x.baseVal.value is 10
PASS rect.y.animVal.value is 10
PASS rect.y.baseVal.value is 10
PASS use.x.animVal.value is 25
PASS use.x.baseVal.value is 0
PASS use.y.animVal.value is 0
PASS use.y.baseVal.value is 0
PASS use.transform.animVal.numberOfItems is 1
PASS use.transform.baseVal.numberOfItems is 0
PASS use.transform.animVal.getItem(0).matrix.e is 0
PASS use.transform.animVal.getItem(0).matrix.f is 125.5
PASS successfullyParsed is true

TEST COMPLETE
